Change Stackdriver stats exporter to use metrics util.
