+++ exporters/stats/stackdriver/build.gradle
@@ -10,6 +10,7 @@ dependencies {
 
     compile project(':opencensus-api'),
             project(':opencensus-contrib-monitored-resource-util'),
+            project(':opencensus-exporter-metrics-util'),
             libraries.google_auth,
             libraries.guava
