+++ exporters/stats/stackdriver/src/test/java/io/opencensus/exporter/stats/stackdriver/StackdriverStatsExporterTest.java
@@ -70,7 +70,7 @@ public class StackdriverStatsExporterTest {
             .setExportInterval(NEG_ONE_SECOND)
             .build();
     thrown.expect(IllegalArgumentException.class);
-    thrown.expectMessage("Duration must be positive");
+    thrown.expectMessage("Export interval must be positive");
     StackdriverStatsExporter.createAndRegister(configuration);
   }
 
@@ -105,7 +105,7 @@ public class StackdriverStatsExporterTest {
   @SuppressWarnings("deprecation")
   public void createWithNegativeDuration() throws IOException {
     thrown.expect(IllegalArgumentException.class);
-    thrown.expectMessage("Duration must be positive");
+    thrown.expectMessage("Export interval must be positive");
     StackdriverStatsExporter.createAndRegisterWithCredentialsAndProjectId(
         GoogleCredentials.newBuilder().build(), PROJECT_ID, NEG_ONE_SECOND);
   }
